backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
celery[redis]>=5.3.0
redis>=5.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
anthropic>=0.40.0
requests>=2.31.0
boto3>=1.34.0
openai-whisper>=20231117
demucs>=4.0.0
pydantic>=2.5.0
